# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |-
    WASAText messaging application API.
    Connect with your friends effortlessly using WASAText! Send and receive messages, 
    whether one-on-one or in groups, all from the convenience of your PC.
  version: "1.0.0"
  contact:
    name: Emanuele Panizzi, Enrico Bassetti
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - BearerAuth: []
paths:
...
  /users/me/username:
    put:
      tags: ["User Management"]
      summary: Update current user's username
      description: Updates the current user's username if the new name is not already taken
      operationId: setMyUserName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Username update request payload
              properties:
                name:
                  type: string
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 3
                  maxLength: 16
                  example: "NewUsername"
                  description: New username to set
              required:
                - name
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty success response
                properties: {}
              example: {}
        '400':
          description: Bad request - invalid username format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/me:
    get:
      tags: ["User Management"]
      summary: Get current user's profile
      description: Retrieve the authenticated user's profile information
      operationId: getMyProfile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userId}:
    get:
      tags: ["User Management"]
      summary: Get user profile by ID
      description: Retrieve a specific user's profile information
      operationId: getUserProfile
      parameters:
        - name: userId
          in: path
          required: true
          description: User identifier
          schema:
            type: string
            pattern: '^[a-zA-Z0-9]+$'
            minLength: 6
            maxLength: 64
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/me/photo:
    put:
      tags: ["User Management"]
      summary: Update current user's profile photo
      description: Upload and set a new profile photo for the authenticated user
      operationId: setMyPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Photo upload form data
              properties:
                photo:
                  type: string
                  format: binary
                  description: Profile photo image file
                  minLength: 1
                  maxLength: 10485760 # 10MB max file size
              required:
                - photo
      responses:
        '200':
          description: Profile photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty success response
                properties: {}
              example: {}
        '400':
          description: Bad request - invalid image format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users:
    get:
      tags: ["User Management"]
      summary: Search for users by username
      description: Search for other WASAText users by username
      operationId: searchUsers
      parameters:
        - name: q
          in: query
          description: Username search query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 50
            pattern: '^[a-zA-Z0-9_-]*$'
      responses:
        '200':
          description: List of matching users
          content:
            application/json:
              schema:
                type: object
                description: Search results containing matched users
                properties:
                  users:
                    type: array
                    description: List of users matching the search criteria
                    items:
                      $ref: '#/components/schemas/User'
                    minItems: 0
                    maxItems: 100
              example:
                users:
                  - id: "user123"
                    username: "Maria"
                    photoUrl: "/photos/user123.jpg"
                  - id: "user456"
                    username: "Mario"
                    photoUrl: "/photos/user456.jpg"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: User identifier obtained from login endpoint
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: Error response containing a message explaining what went wrong
      properties:
        message:
          type: string
          description: Error message
          example: "Username already exists"
      required:
        - message
    User:
      type: object
      description: User profile information
      properties:
        id:
          type: string
          description: User identifier
          example: "user123"
          minLength: 1
          maxLength: 36
          pattern: '^[a-zA-Z0-9_-]+$'
        username:
          type: string
          description: Username
          example: "Maria"
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_-]+$'
        photoUrl:
          type: string
          description: URL to user's profile photo
          example: "/photos/user123.jpg"
          minLength: 1
          maxLength: 255
          pattern: '^/.*$'
      required:
        - id
        - username
...